Remove build logs; fix formatting and UI handling
Remove three stale build_test log files. Update dat_loader to push formatted warning strings directly, use wxstr around std::format where needed and use GetFullPath().utf8_string(), and change hex formatting to {:#x}. Update spr_loader to use utf8_string for file path errors and simplify warnings (remove redundant wxstr/mb_str conversions). Use structured bindings in search_handler for finder.found and use utf8_string for item descriptions. Fix integer-division bug in replace_items_window by casting to double when computing progress percentage.

+++ source/ui/replace_items_window.cpp
@@ -427,7 +427,7 @@ void ReplaceItemsDialog::OnExecuteButtonClicked(wxCommandEvent& WXUNUSED(event))
 		}
 
 		done++;
-		const int value = static_cast<int>((done / items.size()) * 100);
+		const int value = static_cast<int>((static_cast<double>(done) / items.size()) * 100);
 		progress->SetValue(std::clamp<int>(value, 0, 100));
 		list->MarkAsComplete(info, total);
 	}
